Added IPC classes to foundation Miscellaneous int -> size_t warning fixes Removed closures and using boost functions instead

+++ Common/dslIniFile.cpp
@@ -258,8 +258,8 @@ bool IniFile::loadFromString(const string& iniData)
         IniSection* pSection = getSection("");
 
         // These need to be set, we'll restore the original values later.
-        mFlags |= mAutoCreateKeys;
-        mFlags |= mAutoCreateSections;
+        mFlags |= (int) mAutoCreateKeys;
+        mFlags |= (int) mAutoCreateSections;
 
         char* buffer = new char[MAX_LINE_BUFFER_SIZE];
         int lines = 0;
@@ -381,8 +381,8 @@ IniSection* IniFile::loadSection(const string& theSection)
         pSection = getSection("");
 
         // These need to be set, we'll restore the original values later.
-        mFlags |= mAutoCreateKeys;
-        mFlags |= mAutoCreateSections;
+        mFlags |= (int) mAutoCreateKeys;
+        mFlags |= (int) mAutoCreateSections;
 
         while(!bDone)
         {
@@ -826,7 +826,7 @@ bool IniFile::createKey(const string& mKey, const string& mValue, const string& 
     bool bAutoKey = (mFlags & ifAutoCreateKeys) > 0 ? true : false;
     bool bReturn  = false;
 
-    mFlags |= mAutoCreateKeys;
+    mFlags |= (int) mAutoCreateKeys;
 
     bReturn = writeValue(mKey, mValue, mComment, szSection);
     if (!bAutoKey)
@@ -890,22 +890,22 @@ IniSection* IniFile::createSection(const string& Section, const string& Comment,
 }
 
 // Simply returns the number of sections in the list.
-unsigned int IniFile::sectionCount()
+size_t IniFile::sectionCount()
 {
     return mSections.size();
 }
 
 // Returns the total number of keys contained within all the sections.
-unsigned int IniFile::keyCount()
+size_t IniFile::keyCount()
 {
-    int nCounter = 0;
+    size_t nCounter = 0;
     SectionItor s_pos;
     for (s_pos = mSections.begin(); s_pos != mSections.end(); s_pos++)
         nCounter += (*s_pos)->mKeys.size();
     return nCounter;
 }
 
-unsigned int IniFile::keyCount(const string& section)
+size_t IniFile::keyCount(const string& section)
 {
     //Get the section
     IniSection* iniSection = getSection(section);
@@ -1034,9 +1034,9 @@ int IniFile::writeLine(fstream& stream, const char* fmt, ...)
     return nLength;
 }
 
-unsigned int IniFile::getNumberOfSections()
+size_t IniFile::getNumberOfSections()
 {
-    return (int) mSections.size();
+    return mSections.size();
 }
 
 string IniFile::getFileName()
